+++ compiler/hir-lowering/src/lib.rs
@@ -35,6 +35,7 @@ mod items;
 mod operator_traits;
 mod statements;
 mod types;
+mod val_else;
 
 #[cfg(test)]
 mod tests;