+++ compiler/llvm-backend/src/codegen/mod.rs
@@ -8,4 +8,5 @@ pub(crate) mod functions;
 pub(crate) mod panic;
 pub(crate) mod statements;
 pub(crate) mod structs;
+pub(crate) mod val_else;
 pub(crate) use context::CodegenContext;